Separate out postgresql resource type
